+++ app/modules/admin/controllers/UsersController.php
@@ -2,11 +2,11 @@
 
 namespace app\modules\admin\controllers;
 
-use Yii;
-use app\modules\admin\forms\UserForm;
-use app\traits\controllers\FindModelOrFail;
 use app\models\User;
+use app\modules\admin\forms\UserForm;
 use app\modules\admin\models\UserSearch;
+use app\traits\controllers\FindModelOrFail;
+use Yii;
 use yii\filters\VerbFilter;
 
 /**
@@ -15,16 +15,16 @@ use yii\filters\VerbFilter;
 class UsersController extends Controller
 {
 	use FindModelOrFail;
-
+	
 	/**
 	 * @inheritdoc
 	 */
 	public function init()
 	{
 		parent::init();
-		$this->modelClass = UserForm::className();
+		$this->modelClass = UserForm::class;
 	}
-
+	
 	/**
 	 * @inheritdoc
 	 */
@@ -32,14 +32,14 @@ class UsersController extends Controller
 	{
 		return [
 			'verbs' => [
-				'class'   => VerbFilter::className(),
+				'class'   => VerbFilter::class,
 				'actions' => [
 					'delete' => ['POST'],
 				],
 			],
 		];
 	}
-
+	
 	/**
 	 * Lists all User models.
 	 *
@@ -47,15 +47,15 @@ class UsersController extends Controller
 	 */
 	public function actionIndex()
 	{
-		$searchModel  = new UserSearch();
+		$searchModel = new UserSearch();
 		$dataProvider = $searchModel->search(Yii::$app->request->queryParams);
-
+		
 		return $this->render('index', [
 			'searchModel'  => $searchModel,
 			'dataProvider' => $dataProvider,
 		]);
 	}
-
+	
 	/**
 	 * Creates a new User model.
 	 * If creation is successful, the browser will be redirected to the 'view' page.
@@ -65,18 +65,18 @@ class UsersController extends Controller
 	public function actionCreate()
 	{
 		$model = new UserForm();
-
+		
 		$model->on(User::EVENT_BEFORE_INSERT, [$model, 'generateAuthKey']);
-
+		
 		if ($model->load(Yii::$app->request->post()) && $model->save()) {
 			return $this->redirect(['update', 'id' => $model->id]);
 		}
-
+		
 		return $this->render('create', [
 			'model' => $model,
 		]);
 	}
-
+	
 	/**
 	 * Updates an existing User model.
 	 * If update is successful, the browser will be redirected to the 'view' page.
@@ -87,17 +87,20 @@ class UsersController extends Controller
 	 */
 	public function actionUpdate($id)
 	{
+		/**
+		 * @var UserForm $model
+		 */
 		$model = $this->findModel($id);
-
+		
 		if ($model->load(Yii::$app->request->post()) && $model->save()) {
 			return $this->redirect(['update', 'id' => $model->id]);
 		}
-
+		
 		return $this->render('update', [
 			'model' => $model,
 		]);
 	}
-
+	
 	/**
 	 * Deletes an existing User model.
 	 * If deletion is successful, the browser will be redirected to the 'index' page.
@@ -109,8 +112,8 @@ class UsersController extends Controller
 	public function actionDelete($id)
 	{
 		$this->findModel($id)->delete();
-
+		
 		return $this->redirect(['index']);
 	}
-
+	
 }