+++ config/app-console.php
@@ -1,6 +1,7 @@
 <?php
 
 \Yii::setAlias('@app', dirname(__DIR__) . '/app');
+\Yii::setAlias('@webroot', dirname(__DIR__) . '/public');
 
 $db       = require __DIR__ . '/db.php';
 $settings = require __DIR__ . '/settings.php';
@@ -10,7 +11,7 @@ $config = [
 	'basePath'   => dirname(__DIR__) . '/app',
 	'runtimePath'   => dirname(__DIR__) . '/runtime',
 	'vendorPath'   => dirname(__DIR__) . '/vendor',
-	'controllerNamespace' => 'app\\console\\controllers',
+	'controllerNamespace' => 'app\\console',
 	'bootstrap'  => ['log', 'settings'],
 	'aliases'    => [
 		'@config'=> dirname(__DIR__) . '/config',
@@ -20,34 +21,38 @@ $config = [
 	],
 	'controllerMap' => [
 		'migrate' => [
-			'class' => 'app\console\controllers\MigrateController',
+			'class' => \app\modules\base\console\MigrateController::class,
 			'migrationPath' => [
 				'@migrations',
 				'@yii/rbac/migrations',
 				'@vendor/justcoded/yii2-settings/migrations'
 			],
 		],
+		'serve' => [
+			'class'   => \yii\console\controllers\ServeController::class,
+			'docroot' => '@webroot',
+		],
 		'fixture' => [
-			'class' => 'yii\faker\FixtureController',
+			'class' => \yii\faker\FixtureController::class,
 			'namespace' => 'app\fixtures',
 			'templatePath' => '@fixtures/templates',
 			'fixtureDataPath' => '@fixtures/data',
 		],
 		'rbac' => [
-			'class' => 'justcoded\yii2\rbac\commands\RbacController',
+			'class' => \justcoded\yii2\rbac\commands\RbacController::class,
 		],
 	],
 	'components'          => [
 		'authManager' => [
-			'class' => 'justcoded\yii2\rbac\components\DbManager',
+			'class' => \justcoded\yii2\rbac\components\DbManager::class,
 		],
 		'cache' => [
-			'class' => 'yii\caching\FileCache',
+			'class' => \yii\caching\FileCache::class,
 		],
 		'log'   => [
 			'targets' => [
 				[
-					'class'  => 'yii\log\FileTarget',
+					'class'  => \yii\log\FileTarget::class,
 					'levels' => ['error', 'warning'],
 				],
 			],
@@ -58,7 +63,7 @@ $config = [
 ];
 
 if (YII_ENV_DEV) {
-	// configuration adjustments for 'dev' environment
+	// configuration adjustments for 'dev' environment.
 	$config['bootstrap'][]    = 'gii';
 	$config['modules']['gii'] = [
 		'class' => 'yii\gii\Module',