+++ CHANGELOG.md
@@ -3,6 +3,39 @@ CHANGELOG for Yii2 STARTER PROJECT TEMPLATE
 
 *should be replaced with real project changelog later*
 
+v0.9.6
+---------------------
+* Fix register page, controller missed use statement for RegisterForm.
+
+v0.9.5
+---------------------
+* Fix yii rbac/scan for Windows OS (SettingsController generated an error for some reason)
+
+v0.9.4
+---------------------
+* Updated configs to PHP 7 syntax (with ::class call, instead of hard-coded text value)
+
+v0.9.3
+---------------------
+* Fix './yii serve' command.
+* Updated configs to PHP 7 syntax (with ::class call, instead of hard-coded text value)
+
+v0.9.2
+---------------------
+* Fix user migration after code refactoring in v0.9
+
+v0.9.1
+---------------------
+* Fix Codeception tests errors when run through PHPStorm.
+
+v0.9
+---------------------
+* Upgrade app folder structure: moved controllers from web to root, moved commands to console, 
+* Moved migration console/views custom configuration to a separate module.
+* Admin panel: Added user "Roles" fields to edit form.
+* Issue #13: Registration is absent
+* Issue #14: Reset password
+
 v0.8.6
 ---------------------
 * Fixed unit/functional tests